Bootstrap the App using Backbone.Router
To support selecting Movies from the list, we add a SelectionModel to
track which model is selected in the Movies collection. It emits a
'select:none' event when no model is selected. This is a great example
of how easy it is to leverage Backbone.Events to add event support.

When the router triggers the listMovies route, we call #selectModel and
let any listeners react.

In this case, we bound a callback in main.js.