Add a new Movie using Backbone
App.FormView encapsulates Rails-like validation behavior; errors are
shown in-line with the form fields if there is a client-side validation
error. To support this, we mixin #errorMessageFor to Backbone.Model, so
that the validation error messages are defined on the model.

We are not waiting for the create to succeed (we do not use wait: true)!

The new Movie is available on a refresh of the index page.