Show Movie detail with Backbone
We can navigate to a Movie's details without going back to the server.

The SelectionModel emits a 'select' event when we navigate to a Movie.

App.CollectionDetailView manages a set of child views, showing only one
at a time.

We wire up the list and detail views so that they show/hide in
synchrony, for example when we see the list, the detail hides.

And the detail view navigates back to the list without the need for a
server round-trip!