Render individual Movies in the Collection
Presenter wraps a Model. Use it to expose a model's attributes nicely to
Mustache. Add presentation helper methods to the presenter, not the
model.

Like CollectionView, View takes a function to generate the presenter
during rendering. Prefer this kind of configuration to something more static like passing
the class name.