Extract view coordination to controller
Our main.js is filling the role of a controller, so we extract the view
initialization and coordination.

For this simple app, the controller creates the views and listens
for routes (higher level navigation). The views manage themselves with
the help of the SelectionModel.

Note that the Router is just a list of named routes.

In a more complex app, there would be multiple controllers, and each
controller would manage its own views.